move classes into one package and some minor changes
